fixed crash when creating a pdf without text, prevent ocr without selecting columns first. added license headers to source files

+++ textfairy/src/com/renard/ocr/BaseDocumentActivitiy.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2012,2013 Renard Wellnitz.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
 package com.renard.ocr;
 
 import java.io.File;
@@ -121,8 +137,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 
 	@Override
 	public boolean onCreateOptionsMenu(Menu menu) {
-		getSupportMenuInflater().inflate(R.menu.base_document_activity_options,
-				menu);
+		getSupportMenuInflater().inflate(R.menu.base_document_activity_options, menu);
 		return true;
 	}
 
@@ -135,8 +150,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 				int nativePix = data.getIntExtra(EXTRA_NATIVE_PIX, 0);
 				Intent intent = new Intent(this, OCRActivity.class);
 				intent.putExtra(EXTRA_NATIVE_PIX, nativePix);
-				intent.putExtra(OCRActivity.EXTRA_PARENT_DOCUMENT_ID,
-						getParentId());
+				intent.putExtra(OCRActivity.EXTRA_PARENT_DOCUMENT_ID, getParentId());
 				startActivityForResult(intent, REQUEST_CODE_OCR);
 
 				break;
@@ -215,8 +229,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 				options.inPreferredConfig = Bitmap.Config.ARGB_8888;
 				options.inScaled = false;
 				is = getContentResolver().openInputStream(imageUri);
-				final Bitmap bitmap = BitmapFactory.decodeStream(is, null,
-						options);
+				final Bitmap bitmap = BitmapFactory.decodeStream(is, null, options);
 				if (bitmap != null) {
 					final Pix p = ReadFile.readBitmap(bitmap);
 					bitmap.recycle();
@@ -266,34 +279,29 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 			deleteProgressdialog.setMessage(message);
 			deleteProgressdialog.setIndeterminate(false);
 			deleteProgressdialog.setMax(max);
-			deleteProgressdialog
-					.setProgressStyle(ProgressDialog.STYLE_HORIZONTAL);
+			deleteProgressdialog.setProgressStyle(ProgressDialog.STYLE_HORIZONTAL);
 			deleteProgressdialog.setCancelable(false);
 			return deleteProgressdialog;
 		case EDIT_TITLE_DIALOG_ID:
-			View layout = getLayoutInflater().inflate(
-					R.layout.edit_title_dialog, null);
-			final Uri documentUri = Uri.parse(args
-					.getString(DIALOG_ARG_DOCUMENT_URI));
+			View layout = getLayoutInflater().inflate(R.layout.edit_title_dialog, null);
+			final Uri documentUri = Uri.parse(args.getString(DIALOG_ARG_DOCUMENT_URI));
 			final String oldTitle = args.getString(DIALOG_ARG_TITLE);
-			final EditText edit = (EditText) layout
-					.findViewById(R.id.edit_title);
+			final EditText edit = (EditText) layout.findViewById(R.id.edit_title);
 			edit.setText(oldTitle);
 
 			AlertDialog.Builder builder = new Builder(this);
 			builder.setView(layout);
 			builder.setTitle(R.string.edit_dialog_title);
 			builder.setIcon(R.drawable.fairy_showing);
-			builder.setPositiveButton(android.R.string.ok,
-					new OnClickListener() {
+			builder.setPositiveButton(android.R.string.ok, new OnClickListener() {
 
-						@Override
-						public void onClick(DialogInterface dialog, int which) {
-							final String title = edit.getText().toString();
-							saveTitle(title, documentUri);
+				@Override
+				public void onClick(DialogInterface dialog, int which) {
+					final String title = edit.getText().toString();
+					saveTitle(title, documentUri);
 
-						}
-					});
+				}
+			});
 			builder.setNegativeButton(R.string.cancel, new OnClickListener() {
 
 				@Override
@@ -310,11 +318,9 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 	protected void onPrepareDialog(int id, Dialog dialog, Bundle args) {
 		switch (id) {
 		case EDIT_TITLE_DIALOG_ID:
-			final Uri documentUri = Uri.parse(args
-					.getString(DIALOG_ARG_DOCUMENT_URI));
+			final Uri documentUri = Uri.parse(args.getString(DIALOG_ARG_DOCUMENT_URI));
 			final String oldTitle = args.getString(DIALOG_ARG_TITLE);
-			final EditText edit = (EditText) dialog
-					.findViewById(R.id.edit_title);
+			final EditText edit = (EditText) dialog.findViewById(R.id.edit_title);
 			edit.setText(oldTitle);
 			AlertDialog alertDialog = (AlertDialog) dialog;
 			Button okButton = alertDialog.getButton(AlertDialog.BUTTON_NEUTRAL);
@@ -371,8 +377,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 		super.onPrepareDialog(id, dialog, args);
 	}
 
-	protected void askUserForNewTitle(final String oldTitle,
-			final Uri documentUri) {
+	protected void askUserForNewTitle(final String oldTitle, final Uri documentUri) {
 		Bundle bundle = new Bundle(2);
 		bundle.putString(DIALOG_ARG_TITLE, oldTitle);
 		bundle.putString(DIALOG_ARG_DOCUMENT_URI, documentUri.toString());
@@ -385,8 +390,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 			uri = getIntent().getData();
 		}
 		if (uri != null) {
-			SaveDocumentTask saveTask = new SaveDocumentTask(documentUri,
-					newTitle);
+			SaveDocumentTask saveTask = new SaveDocumentTask(documentUri, newTitle);
 			saveTask.execute();
 		}
 
@@ -398,9 +402,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 	 * 
 	 */
 
-	protected class CreatePDFTask extends
-			AsyncTask<Void, Integer, ArrayList<Uri>> implements
-			PDFProgressListener {
+	protected class CreatePDFTask extends AsyncTask<Void, Integer, ArrayList<Uri>> implements PDFProgressListener {
 
 		private Set<Integer> mIds = new HashSet<Integer>();
 		private int mCurrentPageCount;
@@ -416,13 +418,10 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 		public void onNewPage(int pageNumber) {
 			Bundle args = new Bundle(5);
 
-			String progressMsg = getResources().getString(
-					R.string.progress_pfd_creation);
-			progressMsg = String.format(progressMsg, pageNumber,
-					mCurrentPageCount, mCurrentDocumentName);
+			String progressMsg = getResources().getString(R.string.progress_pfd_creation);
+			progressMsg = String.format(progressMsg, pageNumber, mCurrentPageCount, mCurrentDocumentName);
 
-			String title = getResources().getString(
-					R.string.pdf_creation_message);
+			String title = getResources().getString(R.string.pdf_creation_message);
 
 			args.putString(DIALOG_ARG_MESSAGE, title);
 			args.putString(DIALOG_ARG_MESSAGE, progressMsg);
@@ -448,35 +447,24 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 		protected void onPostExecute(ArrayList<Uri> pdfFiles) {
 			dismissDialog(PDF_PROGRESS_DIALOG_ID);
 			if (pdfFiles != null) {
-				Intent shareIntent = new Intent(
-						android.content.Intent.ACTION_SEND_MULTIPLE);
-				shareIntent.putExtra(android.content.Intent.EXTRA_SUBJECT,
-						getText(R.string.share_subject));
+				Intent shareIntent = new Intent(android.content.Intent.ACTION_SEND_MULTIPLE);
+				shareIntent.putExtra(android.content.Intent.EXTRA_SUBJECT, getText(R.string.share_subject));
 				CharSequence seq = Html.fromHtml(mOCRText.toString());
 				shareIntent.putExtra(android.content.Intent.EXTRA_TEXT, seq);
 				// shareIntent.putExtra(android.content.Intent.EXTRA_TEXT,
 				// mHOCRText.toString());
 				shareIntent.setType("application/pdf");
 
-				shareIntent.putParcelableArrayListExtra(Intent.EXTRA_STREAM,
-						pdfFiles);
-				startActivity(Intent.createChooser(shareIntent,
-						getText(R.string.share_chooser_title)));
+				shareIntent.putParcelableArrayListExtra(Intent.EXTRA_STREAM, pdfFiles);
+				startActivity(Intent.createChooser(shareIntent, getText(R.string.share_chooser_title)));
 			} else {
-				Toast.makeText(getApplicationContext(),
-						getText(R.string.error_create_file), Toast.LENGTH_LONG)
-						.show();
+				Toast.makeText(getApplicationContext(), getText(R.string.error_create_file), Toast.LENGTH_LONG).show();
 			}
 		}
 
 		private Pair<File, File> createPDF(File dir, long documentId) {
 
-			Cursor cursor = getContentResolver().query(
-					DocumentContentProvider.CONTENT_URI,
-					null,
-					Columns.PARENT_ID + "=? OR " + Columns.ID + "=?",
-					new String[] { String.valueOf(documentId),
-							String.valueOf(documentId) }, "created ASC");
+			Cursor cursor = getContentResolver().query(DocumentContentProvider.CONTENT_URI, null, Columns.PARENT_ID + "=? OR " + Columns.ID + "=?", new String[] { String.valueOf(documentId), String.valueOf(documentId) }, "created ASC");
 			cursor.moveToFirst();
 
 			int index = cursor.getColumnIndex(Columns.TITLE);
@@ -486,31 +474,34 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 
 			mCurrentDocumentName = fileName;
 			mCurrentPageCount = cursor.getCount();
-			String[] images = new String[cursor.getCount()];
+			String[] images = new String[cursor.getCount()];			
 			String[] hocr = new String[cursor.getCount()];
 			cursor.moveToPosition(-1);
 			while (cursor.moveToNext()) {
 				index = cursor.getColumnIndex(Columns.HOCR_TEXT);
-				hocr[cursor.getPosition()] = cursor.getString(index);
 				index = cursor.getColumnIndex(Columns.PHOTO_PATH);
 				Uri imageUri = Uri.parse(cursor.getString(index));
-				images[cursor.getPosition()] = Util.getPathForUri(
-						BaseDocumentActivitiy.this, imageUri);
+				images[cursor.getPosition()] = Util.getPathForUri(BaseDocumentActivitiy.this, imageUri);
 				index = cursor.getColumnIndex(Columns.OCR_TEXT);
 				final String text = cursor.getString(index);
-				FileWriter writer;
-				try {
-					writer = new FileWriter(outText);
-					writer.write(Html.fromHtml(text).toString());
-					writer.close();
-				} catch (IOException ioException) {
-					if (outText.exists()) {
-						outText.delete();
+				if (text != null && text.length()>0) {
+					hocr[cursor.getPosition()] = cursor.getString(index);
+					FileWriter writer;
+					try {
+						writer = new FileWriter(outText);
+						writer.write(Html.fromHtml(text).toString());
+						writer.close();
+					} catch (IOException ioException) {
+						if (outText.exists()) {
+							outText.delete();
+						}
+						outText = null;
 					}
-					outText = null;
-				}
 
-				mOCRText.append(text);
+					mOCRText.append(text);
+				} else {
+					hocr[cursor.getPosition()] = "";
+				}
 			}
 			cursor.close();
 			Hocr2Pdf pdf = new Hocr2Pdf(this);
@@ -555,8 +546,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 		private final static int RESULT_REMOTE_EXCEPTION = -1;
 		final boolean mFinishActivity;
 
-		public DeleteDocumentTask(Set<Integer> parentDocumentIds,
-				final boolean finishActivityAfterExecution) {
+		public DeleteDocumentTask(Set<Integer> parentDocumentIds, final boolean finishActivityAfterExecution) {
 			mIds.addAll(parentDocumentIds);
 			mFinishActivity = finishActivityAfterExecution;
 		}
@@ -574,9 +564,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 		@Override
 		protected void onPostExecute(Integer result) {
 			if (result == RESULT_REMOTE_EXCEPTION) {
-				Toast.makeText(getApplicationContext(),
-						getText(R.string.delete_error), Toast.LENGTH_LONG)
-						.show();
+				Toast.makeText(getApplicationContext(), getText(R.string.delete_error), Toast.LENGTH_LONG).show();
 			} else if (result > 0) {
 				// String deleteMsg = null;
 				// if (result == 1) {
@@ -595,13 +583,10 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 			}
 		}
 
-		private int deleteDocument(Cursor c, ContentProviderClient client)
-				throws RemoteException {
+		private int deleteDocument(Cursor c, ContentProviderClient client) throws RemoteException {
 			int index = c.getColumnIndex(Columns.ID);
 			int currentId = c.getInt(index);
-			Uri currentDocumentUri = Uri.withAppendedPath(
-					DocumentContentProvider.CONTENT_URI,
-					String.valueOf(currentId));
+			Uri currentDocumentUri = Uri.withAppendedPath(DocumentContentProvider.CONTENT_URI, String.valueOf(currentId));
 			index = c.getColumnIndex(Columns.PHOTO_PATH);
 			String imagePath = c.getString(index);
 			if (imagePath != null) {
@@ -613,21 +598,13 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 		@Override
 		protected Integer doInBackground(Void... params) {
 
-			ContentProviderClient client = getContentResolver()
-					.acquireContentProviderClient(
-							DocumentContentProvider.CONTENT_URI);
+			ContentProviderClient client = getContentResolver().acquireContentProviderClient(DocumentContentProvider.CONTENT_URI);
 
 			int count = 0;
 			int progress = 0;
 			for (Integer id : mIds) {
 				try {
-					Cursor c = client.query(
-							DocumentContentProvider.CONTENT_URI,
-							null,
-							Columns.PARENT_ID + "=? OR " + Columns.ID + "=?",
-							new String[] { String.valueOf(id),
-									String.valueOf(id) }, Columns.PARENT_ID
-									+ " ASC");
+					Cursor c = client.query(DocumentContentProvider.CONTENT_URI, null, Columns.PARENT_ID + "=? OR " + Columns.ID + "=?", new String[] { String.valueOf(id), String.valueOf(id) }, Columns.PARENT_ID + " ASC");
 
 					while (c.moveToNext()) {
 						count += deleteDocument(c, client);
@@ -663,8 +640,7 @@ public abstract class BaseDocumentActivitiy extends MonitoredActivity {
 
 		@Override
 		protected void onPreExecute() {
-			mSaveToast = Toast.makeText(BaseDocumentActivitiy.this,
-					getText(R.string.saving_document), Toast.LENGTH_LONG);
+			mSaveToast = Toast.makeText(BaseDocumentActivitiy.this, getText(R.string.saving_document), Toast.LENGTH_LONG);
 		}
 
 		@Override
